Add changing icons to custom
